[SCB-986] Move shared code of service discovery to spring-boot-common
