+++ java-chassis-dependencies/pom.xml
@@ -1041,6 +1041,11 @@
         <version>1.1.0-SNAPSHOT</version>
       </dependency>
       <dependency>
+        <groupId>org.apache.servicecomb</groupId>
+        <artifactId>spring-boot-common</artifactId>
+        <version>1.1.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
         <groupId>com.google.protobuf</groupId>
         <artifactId>protobuf-java</artifactId>
         <version>3.5.1</version>
