+++ pom.xml
@@ -109,6 +109,7 @@
     <module>archetypes</module>
     <module>java-chassis-dependencies-springboot2</module>
     <module>spring-boot2-starter-parent</module>
+    <module>spring-boot-common</module>
   </modules>
 
   <build>
