+++ spring-boot2-starter-parent/spring-boot2-starter-discovery/pom.xml
@@ -98,5 +98,9 @@
       <groupId>org.apache.servicecomb</groupId>
       <artifactId>spring-boot2-starter</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.apache.servicecomb</groupId>
+      <artifactId>spring-boot-common</artifactId>
+    </dependency>
   </dependencies>
 </project>
